Initial implementation of Vector Tester class
